cub3D

+++ Circolo4/cub3D_refactor/sources/player_controls/rotate.c
@@ -6,22 +6,23 @@
 /*   By: fcardina <fcardina@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 03:21:44 by fcardina          #+#    #+#             */
-/*   Updated: 2024/07/03 17:23:17 by fcardina         ###   ########.fr       */
+/*   Updated: 2024/07/07 14:26:43 by fcardina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
 /* Applies rotation to the player's direction and plane */
-static int	apply_rotation(t_data *game_data, double rotation_speed)
+static int	apply_rotation(t_data *game_data, float rotation_speed)
 {
 	t_player	*player_data;
-	double		temp_x;
-
+	float		temp_x;
 	player_data = &game_data->player;
+	//Rotation of Direction Vector
 	temp_x = player_data->dir_x;
 	player_data->dir_x = player_data->dir_x * cos(rotation_speed) - player_data->dir_y * sin(rotation_speed);
 	player_data->dir_y = temp_x * sin(rotation_speed) + player_data->dir_y * cos(rotation_speed);
+	//Rotation of Plane Vector:
 	temp_x = player_data->plane_vector_x;
 	player_data->plane_vector_x = player_data->plane_vector_x * cos(rotation_speed) - player_data->plane_vector_y * sin(rotation_speed);
 	player_data->plane_vector_y = temp_x * sin(rotation_speed) + player_data->plane_vector_y * cos(rotation_speed);
@@ -29,10 +30,10 @@ static int	apply_rotation(t_data *game_data, double rotation_speed)
 }
 
 /* Rotates the player's direction based on input */
-int	rotate_player_direction(t_data *game_data, double rotation_direction)
+int	rotate_player_direction(t_data *game_data, float rotation_direction)
 {
 	int		rotation_occurred;
-	double	rotation_speed;
+	float	rotation_speed;
 
 	rotation_occurred = 0;
 	rotation_speed = ROTSPEED * rotation_direction;
