+++ Circolo4/cub3D_refactor/sources/validation/validate_position.c
@@ -1,19 +1,19 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   position.c                                         :+:      :+:    :+:   */
+/*   validate_position.c                                :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: fcardina <fcardina@student.42roma.it>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/17 03:21:44 by fcardina          #+#    #+#             */
-/*   Updated: 2024/07/01 20:08:25 by fcardina         ###   ########.fr       */
+/*   Updated: 2024/07/07 14:31:08 by fcardina         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/cub3d.h"
 
 /* Checks if the position is not colliding with a wall */
-static bool	validate_wall_collision(t_data *game_data, double pos_x, double pos_y)
+static bool	validate_wall_collision(t_data *game_data, float pos_x, float pos_y)
 {
 	if (game_data->map[(int)pos_y][(int)pos_x] == '0')
 		return (true);
@@ -21,7 +21,7 @@ static bool	validate_wall_collision(t_data *game_data, double pos_x, double pos_
 }
 
 /* Checks if the position is within the bounds of the map */
-static bool	validate_map_bounds(t_data *game_data, double pos_x, double pos_y)
+static bool	validate_map_bounds(t_data *game_data, float pos_x, float pos_y)
 {
 	if (pos_x < 0.25 || pos_x >= game_data->mapinfo.width - 1.25)
 		return (false);
@@ -31,7 +31,7 @@ static bool	validate_map_bounds(t_data *game_data, double pos_x, double pos_y)
 }
 
 /* Validates the position based on map bounds and wall collision */
-static bool	validate_position(t_data *game_data, double pos_x, double pos_y)
+static bool	validate_position(t_data *game_data, float pos_x, float pos_y)
 {
 	if (validate_map_bounds(game_data, pos_x, pos_y) && validate_wall_collision(game_data, pos_x, pos_y))
 		return (true);
@@ -39,7 +39,7 @@ static bool	validate_position(t_data *game_data, double pos_x, double pos_y)
 }
 
 /* Verifies if the new position is valid for movement */
-int	validate_move(t_data *game_data, double next_x, double next_y)
+int	validate_move(t_data *game_data, float next_x, float next_y)
 {
 	int	movement_occurred;
